@@ -7,22 +7,21 @@
   "author": "shellscape",
   "homepage": "https://github.com/shellscape/prettier-plugin-package",
   "bugs": "https://github.com/shellscape/prettier-plugin-package/issues",
-  "main": "lib/index.js",
+  "type": "module",
+  "exports": "./lib/index.js",
   "engines": {
-    "node": ">=10.13.0"
+    "node": ">=20.19.0"
   },
   "scripts": {
-    "ci:coverage": "nyc npm run ci:test && nyc report --reporter=text-lcov > coverage.lcov",
+    "ci:coverage": "vitest run --coverage && vitest coverage --reporter=lcov --output-file=coverage.lcov",
     "ci:lint": "npm run lint && npm run security",
     "ci:lint:commits": "commitlint --from=${CIRCLE_BRANCH} --to=${CIRCLE_SHA1}",
-    "ci:test": "npm run test -- --verbose",
-    "commitlint": "commitlint",
-    "commitmsg": "commitlint -e $GIT_PARAMS",
+    "ci:test": "npm run test -- --reporter=verbose",
     "lint": "eslint --fix --cache lib test && prettier --write package.json",
     "lint-staged": "lint-staged",
     "prepublishOnly": "npm run lint",
     "security": "npm audit",
-    "test": "ava"
+    "test": "vitest run test/test.js"
   },
   "files": [
     "lib/",
@@ -36,35 +35,22 @@
     "prettier"
   ],
   "peerDependencies": {
-    "prettier": "^2.0.0 || ^3.0.0"
+    "prettier": "^3.0.0"
   },
   "devDependencies": {
     "@commitlint/cli": "^17.6.5",
     "@commitlint/config-conventional": "^17.6.5",
-    "ava": "^5.3.1",
-    "eslint-config-shellscape": "^6.0.1",
+    "@ianvs/prettier-plugin-sort-imports": "^4.5.1",
     "execa": "^4.1.0",
     "lint-staged": "^13.2.2",
-    "nyc": "^15.1.0",
+    "oxlint": "^1.9.0",
     "pre-commit": "^1.2.2",
-    "prettier": "^2.8.8"
-  },
-  "ava": {
-    "files": [
-      "!**/fixtures/**"
-    ]
+    "prettier": "^3.6.2",
+    "vitest": "^3.2.4"
   },
   "lint-staged": {
     "*.js": [
       "eslint --fix"
-    ]
-  },
-  "nyc": {
-    "include": [
-      "lib/*.js"
-    ],
-    "exclude": [
-      "test/"
     ]
   },
   "pre-commit": "lint-staged"
