@@ -8,15 +8,18 @@
   The above copyright notice and this permission notice shall be
   included in all copies or substantial portions of this Source Code Form.
 */
-const { parsers } = require('prettier/parser-babel');
+import babelParsers from "prettier/plugins/babel";
+import estreeParsers from "prettier/plugins/estree";
+import prettier from "prettier";
 
-const { dependencies } = require('./rules/dependencies');
-const { engines } = require('./rules/engines');
-const { files } = require('./rules/files');
-const { scripts } = require('./rules/scripts');
-const { sort } = require('./rules/sort');
+import { dependencies } from "./rules/dependencies.js";
+import { engines } from "./rules/engines.js";
+import { files } from "./rules/files.js";
+import { scripts } from "./rules/scripts.js";
+import { sort } from "./rules/sort.js";
 
-const { 'json-stringify': parser } = parsers;
+const { parsers } = babelParsers;
+const { "json-stringify": parser } = parsers;
 const { parse } = parser;
 const rePkg = /package\.json$/;
 
@@ -30,23 +33,31 @@ const format = (properties) => {
   return props;
 };
 
-module.exports = {
-  name: 'prettier-plugin-package',
+const plugin = {
+  name: "prettier-plugin-package",
   parsers: {
-    'json-stringify': {
+    "json-stringify": {
       ...parser,
-      parse(...args) {
-        const [, , options] = args;
+      async parse(text, options) {
         const { filepath } = options;
-        const ast = parse(...args);
+        const ast = parse(text, options);
 
         if (rePkg.test(filepath)) {
           const { properties, node } = ast;
           ast.properties = format(properties || node.properties);
         }
 
+        const formattedText = await prettier.format(text, {
+          parser: "json-stringify",
+          plugins: [babelParsers, estreeParsers],
+        });
+        console.log("Formatted text:", formattedText);
+
         return ast;
-      }
-    }
-  }
+      },
+      astFormat: "estree-json",
+    },
+  },
 };
+
+export default plugin;
