@@ -1,13 +1,15 @@
 /* eslint-disable import/no-unresolved, sort-keys */
 
-const test = require('ava');
-const prettier = require('prettier');
+import { test, expect } from 'vitest';
+import prettier  from 'prettier';
 
-test('default', (t) => {
+import plugin from '../lib/index.js';
+
+test('default', async () => {
   const options = {
     filepath: 'package.json',
     parser: 'json-stringify',
-    plugins: ['.']
+    plugins: [plugin]
   };
   const fixture = {
     engines: {
@@ -17,7 +19,7 @@ test('default', (t) => {
   };
 
   const input = JSON.stringify(fixture, null, 2);
-  const output = prettier.format(input, options);
+  const output = await prettier.format(input, options);
 
-  t.snapshot(output);
+  expect(output).toMatchSnapshot();
 });
