@@ -37,4 +37,4 @@ const process = (props) =>
     return prop;
   });
 
-module.exports = { dependencies: process };
+export {  process as dependencies };
