@@ -25,4 +25,4 @@ const process = (props) => {
   return props;
 };
 
-module.exports = { engines: process };
+export { process as engines};
