@@ -58,4 +58,4 @@ const process = (props) => {
   return props;
 };
 
-module.exports = { files: process };
+export { process as files };
