@@ -71,4 +71,4 @@ const sort = (props) => {
   return known.concat(unknown);
 };
 
-module.exports = { sort };
+export { sort };
