@@ -23,4 +23,4 @@ const process = (props) => {
   return props;
 };
 
-module.exports = { scripts: process };
+export { process as scripts };
